Refactor basis on PR comments
