+++ src/components/Content/SearchTextField.js
@@ -1,0 +1,25 @@
+import TextField from '@material-ui/core/TextField'
+import {
+    withStyles,
+} from '@material-ui/core/styles';
+
+const SearchTextField = withStyles((theme) => {
+    return {
+        root: {
+            '& .MuiOutlinedInput-root': {
+                '& fieldset': {
+                    borderColor: theme.palette.secondary.dark,
+                },
+                '&:hover fieldset': {
+                    borderColor: theme.palette.primary.contrastText,
+                },
+                '&.Mui-focused fieldset': {
+                    borderColor: theme.palette.primary.contrastText,
+                },
+            },
+        },
+    }
+})(TextField);
+
+
+export default SearchTextField;
